requirements.txt
# Google Cloud Platform
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# Authentication & Configuration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Workspace APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0

# Environment & Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Async Support
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0 